my_flight_portal_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.4.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.0.0.post2
